Migrated presentation to 3 slide version.
ICERM 2012: Reproducibility in Computational and Experimental
Mathematics wants "lightning talks" (such as this one) to be
three slides at most. I'm assuming the three slide limit does not
include title or acknowledgments.

To comply with this goal, I started a branch (three-slide-version)
that will contain the necessary modifications. The first step to
acheiving this goal is commenting out all but the first two slides,
which will be left unmodified (for the time being). The last of my
three allotted slides will be culled from the material in the slides
I commented out, and then I'll delete the commented slides, and merge
the three-slide-version branch into the master branch.